pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xgen-doc2chunk"
version = "0.1.0"
description = "Convert raw documents into AI-understandable context with intelligent text extraction, table detection, and semantic chunking"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "master0419", email = "7slwm7@khu.ac.kr"},
]
maintainers = [
    {name = "master0419", email = "7slwm7@khu.ac.kr"},
]
keywords = [
    "document-processing",
    "pdf",
    "ocr",
    "text-extraction",
    "chunking",
    "ai",
    "llm",
    "langchain",
    "hwp",
    "docx",
    "xlsx",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "orjson==3.11.5",
    "psutil==7.0.0",
    "pandas==2.3.3",
    "langchain==1.2.3",
    "langchain-aws==1.2.0",
    "langchain-community==0.4.1",
    "langchain-core==1.2.6",
    "langchain-openai==1.1.7",
    "langchain-anthropic==1.3.1",
    "langchain-google-genai==4.1.3",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.5",
    "langsmith==0.6.2",
    "beautifulsoup4==4.14.3",
    "docx2pdf==0.1.8",
    "python-docx==1.2.0",
    "python-pptx==1.0.2",
    "pymupdf==1.26.5",
    "pdfplumber==0.11.5",
    "openpyxl==3.1.5",
    "xlrd==2.0.2",
    "olefile==0.47",
    "pyhwp==0.1b15",
    "chardet==5.2.0",
    "pi-heif==1.1.1",
    "pdfminer-six==20231228",
    "pdf2image==1.17.0",
    "pytesseract==0.3.13",
    "striprtf==0.0.29",
    "cachetools==6.2.4",
]

[project.urls]
Homepage = "https://github.com/master0419/doc2chunk"
Documentation = "https://github.com/master0419/doc2chunk#readme"
Repository = "https://github.com/master0419/doc2chunk.git"
Issues = "https://github.com/master0419/doc2chunk/issues"
Changelog = "https://github.com/master0419/doc2chunk/releases"

[tool.hatch.build.targets.wheel]
packages = ["xgen-doc2chunk"]

[tool.hatch.build.targets.sdist]
include = [
    "xgen-doc2chunk/",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]
exclude = [
    "/.git",
    "/.venv",
    "/__pycache__",
    "*.pyc",
    "test/",
    "temp/",
    "libs/",
]
